test(ui): share tree view test harness
